requirements.txt
Django==1.10
django-bootstrap3==7.0.1
django-fontawesome==0.2.6
psycopg2==2.6.2
travis==0.0.2
djangorestframework==3.3.3
gunicorn==18.0
dj-static==0.0.6
static3==0.5.1
django-cors-headers==1.3.0
